travis: Add travis file to regularly test compilations.
Can not yet run on docker-based-infrastructure since Qt 5 is needed.
Since Qt 5.0 is the only available version for now for Ubuntu 12.04,
clang can not be tested (yet).
As such, we test compilation with g++-5.
Most pieces to move to the new travis-infrastructure (once Qt is whitelisted)
are already in place, though.